Moved map to separate component
Google Map is now in separate container "MapContainer"
