+++ FrontEnd/debug.log
@@ -11,3 +11,5 @@
 [0228/181413.667:ERROR:crash_report_database_win.cc(428)] unexpected header
 [0228/181631.676:ERROR:crash_report_database_win.cc(428)] unexpected header
 [0306/173622.685:ERROR:crash_report_database_win.cc(428)] unexpected header
+[0402/153143.061:ERROR:crash_report_database_win.cc(428)] unexpected header
+[0402/153625.418:ERROR:crash_report_database_win.cc(428)] unexpected header
